Splits ClipList related types and functions out to their own files
