+++ PlayRho/Collision/Collision.cpp
@@ -22,8 +22,6 @@
 #include <PlayRho/Collision/Collision.hpp>
 #include <PlayRho/Collision/Manifold.hpp>
 
-#include <cmath>
-
 namespace playrho {
 namespace d2 {
 
@@ -62,49 +60,6 @@ PointStates GetPointStates(const Manifold& manifold1, const Manifold& manifold2)
     }
     
     return retval;
-}
-
-ClipList ClipSegmentToLine(const ClipList& vIn, const UnitVec& normal, Length offset,
-                           ContactFeature::Index indexA)
-{
-    ClipList vOut;
-
-    if (size(vIn) == 2) // must have two points (for a segment)
-    {
-        // Use Sutherland-Hodgman clipping:
-        //   (https://en.wikipedia.org/wiki/Sutherland%E2%80%93Hodgman_algorithm ).
-
-        // Calculate the distance of end points to the line
-        const auto distance0 = Dot(normal, vIn[0].v) - offset;
-        const auto distance1 = Dot(normal, vIn[1].v) - offset;
-
-        // If the points are behind the plane...
-        // Ideally they are. Then we get face-vertex contact features which are simpler to
-        // calculate. Note that it also helps to avoid changing the contact feature from the
-        // given clip vertices. So the code here also accepts distances that are just slightly
-        // over zero.
-        if (distance0 <= 0_m || AlmostZero(StripUnit(distance0)))
-        {
-            vOut.push_back(vIn[0]);
-        }
-        if (distance1 <= 0_m || AlmostZero(StripUnit(distance1)))
-        {
-            vOut.push_back(vIn[1]);
-        }
-
-        // If we didn't already find two points & the points are on different sides of the plane...
-        if (size(vOut) < 2 && signbit(StripUnit(distance0)) != signbit(StripUnit(distance1)))
-        {
-            // Neither distance0 nor distance1 is 0 and either one or the other is negative (but not both).
-            // Find intersection point of edge and plane
-            // Vertex A is hitting edge B.
-            const auto interp = distance0 / (distance0 - distance1);
-            const auto vertex = vIn[0].v + (vIn[1].v - vIn[0].v) * interp;
-            vOut.push_back(ClipVertex{vertex, GetVertexFaceContactFeature(indexA, vIn[0].cf.indexB)});
-        }
-    }
-
-    return vOut;
 }
 
 } // namespace d2
