+++ PlayRho/Collision/Collision.hpp
@@ -23,61 +23,37 @@
 #define PLAYRHO_COLLISION_COLLISION_HPP
 
 /// @file
-/// Structures and functions used for computing contact points, distance
-/// queries, and TOI queries.
+/// Structures and functions used for computing before and after like point oriented
+/// collision response states.
 
-#include <PlayRho/Common/Math.hpp>
-#include <PlayRho/Common/ArrayList.hpp>
-#include <PlayRho/Collision/ContactFeature.hpp>
+#include <PlayRho/Common/Settings.hpp>
+
 #include <PlayRho/Collision/PointState.hpp>
 
-namespace playrho {
+namespace playrho::d2 {
+
+class Manifold;
 
 /// @brief Point states.
 /// @details The states pertain to the transition from an old manifold to a new manifold.
 ///   So state 1 is either persist or remove while state 2 is either add or persist.
+/// @see Manifold, GetPointStates.
 struct PointStates
 {
     /// @brief State 1.
+    /// @details States for the first manifold.
     PointState state1[MaxManifoldPoints] = {PointState::Null, PointState::Null};
     
     /// @brief State 2.
+    /// @details States after the second manifold.
     PointState state2[MaxManifoldPoints] = {PointState::Null, PointState::Null};
 };
 
-namespace d2 {
-
-class Manifold;
-
-/// @brief Computes the point states given two manifolds.
+/// @brief Computes the _before_ and _after_ like point states given two manifolds.
+/// @note This can be useful for analyzing collision responses like in the world's pre-solve
+///   contact event listener.
 PointStates GetPointStates(const Manifold& manifold1, const Manifold& manifold2) noexcept;
 
-/// @brief Clip vertex.
-/// @details Used for computing contact manifolds.
-/// @note This data structure is 12-bytes large (on at least one 64-bit platform).
-struct ClipVertex
-{
-    Length2 v; ///< Vertex of edge or polygon. 8-bytes.
-    ContactFeature cf; ///< Contact feature information. 4-bytes.
-};
-
-/// @brief Clip list for <code>ClipSegmentToLine</code>.
-/// @see ClipSegmentToLine.
-/// @note This data structure is at least 24-bytes large.
-using ClipList = ArrayList<ClipVertex, MaxManifoldPoints>;
-
-/// Clipping for contact manifolds.
-/// @details This returns an array of points from the given line that are inside of the plane as
-///   defined by a given normal and offset.
-/// @param vIn Clip list of two points defining the line.
-/// @param normal Normal of the plane with which to determine intersection.
-/// @param offset Offset of the plane with which to determine intersection.
-/// @param indexA Index of vertex A.
-/// @return List of zero one or two clip points.
-ClipList ClipSegmentToLine(const ClipList& vIn, const UnitVec& normal, Length offset,
-                           ContactFeature::Index indexA);
-
-} // namespace d2
-} // namespace playrho
+}
 
 #endif // PLAYRHO_COLLISION_COLLISION_HPP
