+++ PlayRho/Dynamics/Contacts/ContactSolver.cpp
@@ -19,7 +19,6 @@
  * 3. This notice may not be removed or altered from any source distribution.
  */
 
-#include <PlayRho/Collision/Collision.hpp>
 #include <PlayRho/Collision/WorldManifold.hpp>
 #include <PlayRho/Dynamics/Contacts/ContactSolver.hpp>
 #include <PlayRho/Dynamics/Contacts/ConstraintSolverConf.hpp>
